requirements.txt
# Summary: Lists Python dependencies required to run and test OrchestrAI.
fastapi==0.135.1
pydantic==2.12.5
requests==2.32.5
python-dotenv==1.2.2
pyyaml==6.0.3